+++ tests/TestCase.php
@@ -23,6 +23,7 @@ abstract class TestCase extends Orchestra
 
     protected function defineEnvironment($app): void
     {
+        $app['config']->set('captchaapi.enabled', true);
         $app['config']->set('captchaapi.site_key', 'test_site_key');
         $app['config']->set('captchaapi.secret_keys', ['test_secret_key']);
         $app['config']->set('captchaapi.replay_protection', true);